Refactor: 영업기회
영업기회 상태 패키지 변경
