+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/aggregate/entity/SalesOpp.java
@@ -1,7 +1,6 @@
 package com.cineverse.erpc.salesopp.opportunity.aggregate.entity;
 
 import com.cineverse.erpc.employee.aggregate.Employee;
-import com.cineverse.erpc.salesopp.opportunity.aggregate.vo.SalesOppStatus;
 import jakarta.persistence.*;
 import lombok.*;
