+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/aggregate/entity/SalesOppStatus.java
@@ -1,4 +1,4 @@
-package com.cineverse.erpc.salesopp.opportunity.aggregate.vo;
+package com.cineverse.erpc.salesopp.opportunity.aggregate.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
